(feat) PureMMForm, load forms based on strategyName prop
